requirements.txt
# Build Tools
setuptools>=70.0.0
wheel>=0.43.0

# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.11.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Data Processing
numpy==2.0.0
pandas==2.2.3
requests==2.31.0
python-dateutil==2.8.2
